services/sound-classifier/requirements.txt
fastapi
uvicorn[standard]
webrtcvad-wheels
numpy
tensorflow-cpu